+++ src/model/ControladorDeTransito.cpp
@@ -10,6 +10,7 @@ ControladorDeTransito::ControladorDeTransito(Connection &connection)
     atualizarListas();
 }
 
+
 void ControladorDeTransito::cadastrarCidade() {
     std::string nome;
     cout << "Digite o nome da cidade: ";
@@ -28,6 +29,7 @@ void ControladorDeTransito::cadastrarCidade() {
     atualizarListas();
     delete cidade;
 }
+
 
 void ControladorDeTransito::cadastrarTrajeto() {
     std::string nomeOrigem;
@@ -88,6 +90,7 @@ void ControladorDeTransito::cadastrarTrajeto() {
     atualizarListas();
     delete trajeto;
 }
+
 
 void ControladorDeTransito::cadastrarTransporte(){
     std::string nomeTransporte;
@@ -171,6 +174,7 @@ void ControladorDeTransito::cadastrarTransporte(){
     delete transporte;
 }
 
+
 void ControladorDeTransito::cadastrarPassageiro() {
     string nome;
     string cpf;
@@ -224,257 +228,7 @@ void ControladorDeTransito::cadastrarPassageiro() {
 
 }
 
-void ControladorDeTransito::cadastrarViagem() {
-    std::string nomeTransporte;
-    std::vector<std::string> nomesPassageiros;
-    std::string nomeOrigem;
-    std::string nomeDestino;
-    
-    vector<Trajeto*> melhorTrajeto;
-    int tamanho = static_cast<int>(cidades.size() + 1);
-    Grafo* grafo = new Grafo(tamanho);
 
-    cout << "Digite a cidade de origem: ";
-    getline(cin, nomeOrigem);
-
-    nomeOrigem = para_minusculo(nomeOrigem);
-
-    if (nomeOrigem == "emtrânsito"){
-        cout << "Cidade de origem não pode ser 'EmTrânsito'" << endl;
-        return;
-    }
-
-    Cidade* cidadeOrigem = cidadeDAO.findByNome(nomeOrigem);
-    if (cidadeOrigem == nullptr) {
-        cout << "Cidade não cadastrada no banco de dados" << endl;
-        return;
-    }
-
-    cout << "Digite a cidade de destino: ";
-    getline(cin, nomeDestino);
-
-    nomeDestino = para_minusculo(nomeDestino);
-
-    Cidade* cidadeDestino = cidadeDAO.findByNome(nomeDestino);
-    if (cidadeDestino == nullptr) {
-        cout << "Cidade não cadastrada no banco de dados!" << endl;
-        return;
-    }
-
-    cout << "Digite o nome do transporte: ";
-    getline(cin, nomeTransporte);
-
-    Transporte* transporte = transporteDAO.findByNome(nomeTransporte);
-    if (transporte == nullptr) {
-        cout << "Transporte não cadastrado no banco de dados!" << endl;
-        return;
-    }
-
-    int capacidade_transporte = transporte->getCapacidade();
-    string cpf;
-    vector<Passageiro*> passageiros;
-    char sim_ou_nao;
-    for(int i = 0; i < capacidade_transporte; i++){
-        cout << "Digite o cpf do passageiro: ";
-        getline(cin, cpf);
-        Passageiro* passageiro = passageiroDAO.findByCpf(cpf);
-        if (passageiro == nullptr) {
-        cout << "Passageiro não cadastrada no banco de dados!" << endl;
-        return;
-        }
-        passageiros.push_back(passageiro);
-
-        cout << "Quer mais passageiros? (S/N)";
-        cin >> sim_ou_nao;
-        if(sim_ou_nao == 'N' || sim_ou_nao == 'n'){
-            cin.ignore();
-            break;
-        }
-        cin.ignore();
-        if(i == capacidade_transporte - 1){
-            cout << "Quantidade máxima de passageiros alcançada!" << endl;
-        }
-    }
-
-
-    grafo->adicionarAresta(trajetos);
-    melhorTrajeto = grafo->dfs(cidadeOrigem->getId(), cidadeDestino->getId(), transporte->getTipo(), trajetos); // Corrigir chamada para incluir tipo
-    
-    if (melhorTrajeto.empty())
-    {
-        cout << "Nenhuma rota encontrada! " << endl;
-        delete cidadeOrigem;
-        delete cidadeDestino;
-        delete transporte;
-        return;
-    }
-    
-    
-    Viagem* viagem = new Viagem(transporte, passageiros, melhorTrajeto, cidadeOrigem, cidadeDestino, 0, 0);
-
-    viagemDAO.create(*viagem);
-    cout << "Viagem cadastrada com sucesso!" << endl;
-
-    atualizarListas();
-    
-    delete viagem;
-    delete cidadeOrigem;
-    delete cidadeDestino;
-    delete transporte;
-    for (size_t i = 0; i < melhorTrajeto.size(); i++)
-    {
-        delete melhorTrajeto[i];
-    }
-    
-}
-
-void ControladorDeTransito::verificarRotas(){
-    string input;
-    int opcao;
-    int id;
-
-    do
-    {
-        //oi
-        cout << "VIAGENS CADASTRADAS" << endl;
-        for (size_t i = 0; i < viagens.size(); i++)
-        {
-            cout << "Viagem #" << i+1 << " - Código: " << viagens[i]->getId() << endl;
-        }
-        
-        cout << "Digite o código da viagem que deseja verificar os trajetos: ";
-        getline(cin, input);
-
-        try
-        {
-            opcao = stoi(input);
-        }
-        catch(const std::exception& e)
-        {
-            cout << "Opção inválida" << endl;
-            break;
-        }
-
-        Viagem* viagem = viagemDAO.findById(opcao);
-
-        if (viagem == nullptr)
-        {
-            cout << "Nenhuma viagem encontrada!" << endl;
-            break;
-        }
-        
-
-        vector<Trajeto*> trajetos = viagem->getTrajetos();
-
-        cout << "A melhor rota da viagem #" << viagem->getId() << " tem os seguintes trajetos: " << endl;
-        for (size_t i = 0; i < trajetos.size(); i++)
-        {
-            cout << "Trajeto #" << i+1 << " - Cidade Origem: " << trajetos[i]->getOrigem()->getNome() << " - Cidade Destino: " << 
-            trajetos[i]->getDestino()->getNome() << endl;
-        }
-
-        cout << "Digite 1 para continuar verificando: ";
-        getline(cin, input);
-
-        try
-        {
-            opcao = stoi(input);
-        }
-        catch(const std::exception& e)
-        {
-            cout<< "Opcao invalida!" << endl;
-            break;
-        }
-        
-    } while (opcao == 1);
-}
-std::chrono::system_clock::time_point ControladorDeTransito::getCurrentTimePoint() {
-    return std::chrono::system_clock::now();
-}
-
-void ControladorDeTransito::iniciarViagem(){
-    string input;
-    int opcao;
-    
-    cout << "=======LISTA DE VIAGEMS======="<<endl;
-    for (size_t i = 0; i < viagens.size(); i++)
-    {
-        cout << "Viagem #" << i+1 << " - Código: " << viagens[i]->getId() << " - Cidade de Origem: " << viagens[i]->getOrigem()->getNome()
-        << " - Cidade Destino: " << viagens[i]->getDestino()->getNome() << endl;
-    }
-
-    cout << "Digite o código da cidade que deseja iniciar: ";
-    getline(cin, input);
-
-    try{
-        opcao = stoi(input);
-    }
-    catch(const std::exception& e)
-    {
-        cout << "Opcao invalida." << endl;
-        return;
-    }
-
-    Viagem* viagem = viagemDAO.findById(opcao);
-    if (viagem = nullptr){
-        cout << "Viagem não encontrada!" << endl;
-        return;
-    }
-
-    if (viagem->getTransporte()->getLocalAtual() != viagem->getOrigem()){
-        cout << "Transporte: " << viagem->getTransporte()->getNome() << " não está na cidade de origem da viagem." << endl;
-    }
-
-    for (size_t i = 0; i < viagem->getPassageiros().size(); i++){
-        if (viagem->getPassageiros()[i]->getLocalAtual() != viagem->getOrigem()){
-            cout << "Passageiro: " << viagem->getPassageiros()[i]->getNome() << " não está na cidade de origem da viagem." << endl;
-            cout << "Não é possível iniciar viagem!" << endl;
-            return;
-        }
-    }
-
-    //Iniciar mudança de status
-    for (size_t i = 0; i < transportes.size(); i++)
-    {
-        if (transportes[i] == viagem->getTransporte())
-        {
-            transportes[i]->setLocalAtual(cidadeDAO.findById(1));
-            break;
-        }
-        
-    }
-
-    bool encontrou = false;
-    for (size_t i = 0; i < viagem->getPassageiros().size(); i++)
-    {
-        for (size_t j = 0; j < passageiros.size(); j++)
-        {
-            if (viagem->getPassageiros()[i] == passageiros[j])
-            {
-                passageiros[i]->setLocalAtual(cidadeDAO.findById(1));
-                encontrou = true;
-                break;
-            }
-            
-        }
-        if (encontrou)
-        {
-            break;
-        }
-    }
-
-    for (size_t i = 0; i < viagens.size(); i++)
-    {
-        if (viagens[i]->getId() == opcao)
-        {
-            viagens[i]->setStatusViagem(2);
-            break;
-        }
-        
-    }
-    
-    
-}
 
 std::string ControladorDeTransito::timePointToString(const std::chrono::system_clock::time_point& timePoint) {
     std::time_t now_time = std::chrono::system_clock::to_time_t(timePoint);
@@ -484,6 +238,8 @@ std::string ControladorDeTransito::timePointToString(const std::chrono::system_c
     oss << std::put_time(local_time, "%Y-%m-%d %H:%M:%S");
     return oss.str();
 }
+
+
 
 void ControladorDeTransito::avancarHoras(std::chrono::system_clock::time_point& timePoint) {
     // Define a duração de uma hora
@@ -497,96 +253,3 @@ void ControladorDeTransito::avancarHoras(std::chrono::system_clock::time_point& 
 void ControladorDeTransito::relatarEstado() {
 }
 
-void ControladorDeTransito::gerarRelatorios(int tipo) {
-    switch (tipo) {
-        case 1: {
-            std::vector<PassageiroDTO*> passageirosDTO = consultasDAO.relatorioLocalidadePassagerios();
-            for (int i = 0; i < passageirosDTO.size(); i++) {
-                PassageiroDTO* passageiroDTO = passageirosDTO[i];
-                std::cout << passageiroDTO->toString();
-            }
-            break;
-        }
-        case 2: {
-            std::vector<TransporteDTO*> transportesDTO = consultasDAO.relatorioLocalidadeTransportes();
-            for (int i = 0; i < transportesDTO.size(); i++) {
-                TransporteDTO* transporteDTO = transportesDTO[i];
-                std::cout << transporteDTO->toString();
-            }
-            break;
-        }
-        case 3: {
-            std::vector<ViagemDTO*> viagensDTO = consultasDAO.relatorioViagensEmAndamento();
-            for (int i = 0; i < viagensDTO.size(); i++) {
-                ViagemDTO* viagemDTO = viagensDTO[i];
-                std::cout << viagemDTO->toString();
-            }
-            break;
-        }
-        case 4: {
-            std::vector<CidadeDTO*> cidadesDTO = consultasDAO.relatorioCidadesMaisVisitadas();
-            for (int i = 0; i < cidadesDTO.size(); i++) {
-                CidadeDTO* cidadeDTO = cidadesDTO[i];
-                std::cout << cidadeDTO->toString();
-            }
-            break;
-        }
-        case 0:
-            break;
-        default:
-            std::cout << "Opção inválida no Consultas e Relatórios. Tente novamente." << std::endl;
-            break;
-    }
-}
-
-bool ControladorDeTransito::validarCidade(Cidade* cidade) {
-    for(int i = 0; i < cidades.size(); i++) {
-        Cidade* cidadeBanco = cidades[i];
-        if(cidadeBanco->getNome() == cidade->getNome()){
-            return false;
-        }
-    }
-    return true;
-}
-
-void ControladorDeTransito::atualizarListas() {
-    cidades = cidadeDAO.findAll();
-    transportes = transporteDAO.findAll();
-    trajetos = trajetoDAO.findAll();
-    passageiros = passageiroDAO.findAll();
-    viagens = viagemDAO.findAll();
-}
-
-bool ControladorDeTransito::validarPassageiro(Passageiro *passageiro) {
-    for (int i = 0; i < passageiros.size(); i++) {
-        if (passageiro == passageiros[i]) {
-            return false;
-        }
-    }
-    return true;
-}
-
-bool ControladorDeTransito::validarTransporte(Transporte* transporte) {
-    for (int i = 0; i < transportes.size(); i++) {
-        if (transportes[i]->getNome() == transporte->getNome()) {
-            return false;
-        }
-    }
-    return true;
-}
-
-bool ControladorDeTransito::validarTrajeto(Trajeto* trajeto) {
-    for (int i = 0; i < trajetos.size(); i++) {
-        if (trajetos[i]->getOrigem()->getId() == trajeto->getOrigem()->getId() && trajetos[i]->getDestino()->getId() == trajeto->getDestino()->getId() && trajetos[i]->getTipo() == trajeto->getTipo() && trajetos[i]->getDistancia() == trajeto->getDistancia()) {
-            return false;
-        }
-    }
-    return true;
-}
-
-std::string ControladorDeTransito::para_minusculo(const std::string& nome) {
-    std::string resultado = nome;
-    std::transform(resultado.begin(), resultado.end(), resultado.begin(),
-                   [](unsigned char c){ return std::tolower(c); });
-    return resultado;
-}
