+++ src/model/ControladorRelatorios.cpp
@@ -1,0 +1,48 @@
+#include "../../include/model/ControladorDeTransisto.hpp"
+#include <iostream>
+#include <algorithm> 
+#include <cctype> 
+
+using namespace std;
+
+void ControladorDeTransito::gerarRelatorios(int tipo) {
+    switch (tipo) {
+        case 1: {
+            std::vector<PassageiroDTO*> passageirosDTO = consultasDAO.relatorioLocalidadePassagerios();
+            for (int i = 0; i < passageirosDTO.size(); i++) {
+                PassageiroDTO* passageiroDTO = passageirosDTO[i];
+                std::cout << passageiroDTO->toString();
+            }
+            break;
+        }
+        case 2: {
+            std::vector<TransporteDTO*> transportesDTO = consultasDAO.relatorioLocalidadeTransportes();
+            for (int i = 0; i < transportesDTO.size(); i++) {
+                TransporteDTO* transporteDTO = transportesDTO[i];
+                std::cout << transporteDTO->toString();
+            }
+            break;
+        }
+        case 3: {
+            std::vector<ViagemDTO*> viagensDTO = consultasDAO.relatorioViagensEmAndamento();
+            for (int i = 0; i < viagensDTO.size(); i++) {
+                ViagemDTO* viagemDTO = viagensDTO[i];
+                std::cout << viagemDTO->toString();
+            }
+            break;
+        }
+        case 4: {
+            std::vector<CidadeDTO*> cidadesDTO = consultasDAO.relatorioCidadesMaisVisitadas();
+            for (int i = 0; i < cidadesDTO.size(); i++) {
+                CidadeDTO* cidadeDTO = cidadesDTO[i];
+                std::cout << cidadeDTO->toString();
+            }
+            break;
+        }
+        case 0:
+            break;
+        default:
+            std::cout << "Opção inválida no Consultas e Relatórios. Tente novamente." << std::endl;
+            break;
+    }
+}
